refactor: 상태 아이콘 로직 enum으로 분리 및 DesignSystem 의존성 모델로 이동
